Commit on 2 Jul

+++ Pointwise.SqlDataAccess/SqlRepositories/SQLUserRoleRepository.cs
@@ -45,21 +45,27 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
             return insertedRows.Select(x => x.ToDomainEntity()).AsEnumerable();
         }
 
-        public void SoftDelete(int id)
+        public bool SoftDelete(int id)
         {
             var sEntity = context.UserRoles.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
+
             sEntity.IsDeleted = true;
             context.SaveChanges();
+            return true;
         }
 
-        public void UndoSoftDelete(int id)
+        public bool UndoSoftDelete(int id)
         {
             var sEntity = context.UserRoles.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
+
             sEntity.IsDeleted = false;
             context.SaveChanges();
+            return true;
         }
 
-        public void SoftDeleteRange(IEnumerable<Domain.Models.UserRole> entities)
+        public bool SoftDeleteRange(IEnumerable<Domain.Models.UserRole> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             foreach (var entity in sEntities)
@@ -67,21 +73,26 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
                 entity.IsDeleted = true;
             }
             context.SaveChanges();
+            return true;
         }
 
-        public void Delete(int id)
+        public bool Delete(int id)
         {
             var sEntity = context.UserRoles.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
+
             context.UserRoles.Remove(sEntity);
             context.SaveChanges();
+            return true;
         }
 
-        public void DeleteRange(IEnumerable<Domain.Models.UserRole> entities)
+        public bool DeleteRange(IEnumerable<Domain.Models.UserRole> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             context.UserRoles.RemoveRange(sEntities);
 
             context.SaveChanges();
+            return true;
         }
 
         public IUserRole Update(Domain.Models.UserRole entity)
