+++ Pointwise.SqlDataAccess/SqlRepositories/SqlCategoryRepository.cs
@@ -47,22 +47,25 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
             return insertedRows.Select(x => x.ToDomainEntity()).AsEnumerable();
         }
 
-        public void SoftDelete(int id)
+        public bool SoftDelete(int id)
         {
             var sEntity = context.Categories.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
             sEntity.IsDeleted = true;
             context.SaveChanges();
-
+            return true;
         }
 
-        public void UndoSoftDelete(int id)
+        public bool UndoSoftDelete(int id)
         {
             var sEntity = context.Categories.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
             sEntity.IsDeleted = false;
             context.SaveChanges();
+            return true;
         }
 
-        public void SoftDeleteRange(IEnumerable<Domain.Models.Category> entities)
+        public bool SoftDeleteRange(IEnumerable<Domain.Models.Category> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             foreach(var entity in sEntities)
@@ -70,21 +73,27 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
                 entity.IsDeleted = true;
             }
             context.SaveChanges();
+            return true;
         }
 
-        public void Delete(int id)
+        public bool Delete(int id)
         {
             var sEntity = context.Categories.SingleOrDefault(x => x.Id == id);
+
+            if (sEntity == null) return false;
+
             context.Categories.Remove(sEntity);
             context.SaveChanges();
+            return true;
         }
 
-        public void DeleteRange(IEnumerable<Domain.Models.Category> entities)
+        public bool DeleteRange(IEnumerable<Domain.Models.Category> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             context.Categories.RemoveRange(sEntities);
             
             context.SaveChanges();
+            return true;
         }
 
         public ICategory Update(Domain.Models.Category entity)
