+++ Pointwise.SqlDataAccess/SqlRepositories/SqlSourceRepository.cs
@@ -65,21 +65,25 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
             return insertedRows.Select(x => x.ToDomainEntity()).AsEnumerable();
         }
 
-        public void SoftDelete(int id)
+        public bool SoftDelete(int id)
         {
             var sEntity = context.Sources.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
             sEntity.IsDeleted = true;
             context.SaveChanges();
+            return true;
         }
 
-        public void UndoSoftDelete(int id)
+        public bool UndoSoftDelete(int id)
         {
             var sEntity = context.Sources.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
             sEntity.IsDeleted = false;
             context.SaveChanges();
+            return true;
         }
 
-        public void SoftDeleteRange(IEnumerable<Source> entities)
+        public bool SoftDeleteRange(IEnumerable<Source> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             foreach(var source in sEntities)
@@ -87,20 +91,25 @@ namespace Pointwise.SqlDataAccess.SqlRepositories
                 source.IsDeleted = true;
             }
             context.SaveChanges();
+            return true;
         }
 
-        public void Delete(int id)
+        public bool Delete(int id)
         {
             var sEntity = context.Sources.SingleOrDefault(x => x.Id == id);
+            if (sEntity == null) return false;
+
             context.Sources.Remove(sEntity);
             context.SaveChanges();
+            return true;
         }
 
-        public void DeleteRange(IEnumerable<Source> entities)
+        public bool DeleteRange(IEnumerable<Source> entities)
         {
             var sEntities = entities.Select(x => x.ToPersistentEntity()).AsEnumerable();
             context.Sources.RemoveRange(sEntities);
             context.SaveChanges();
+            return true;
         }
 
         public ISource Update(Source entity)
